RcppArmadillo 0.2.16 with Armadillo 1.1.8

+++ inst/include/armadillo_bits/format_wrap.hpp
@@ -143,7 +143,7 @@ namespace arma_boost
           buffer = new char[buffer_size];
           }
         
-        required_size = arma_snprintf(buffer, buffer_size, X.A.A.c_str(), X.B);
+        required_size = arma_snprintf(buffer, size_t(buffer_size), X.A.A.c_str(), X.B);
         
         if(required_size < buffer_size)
           {
@@ -195,7 +195,7 @@ namespace arma_boost
           buffer = new char[buffer_size];
           }
         
-        required_size = arma_snprintf(buffer, buffer_size, X.A.A.A.c_str(), X.A.B, X.B);
+        required_size = arma_snprintf(buffer, size_t(buffer_size), X.A.A.A.c_str(), X.A.B, X.B);
         
         if(required_size < buffer_size)
           {
@@ -247,7 +247,7 @@ namespace arma_boost
           buffer = new char[buffer_size];
           }
         
-        required_size = arma_snprintf(buffer, buffer_size, X.A.A.A.A.c_str(), X.A.A.B, X.A.B, X.B);
+        required_size = arma_snprintf(buffer, size_t(buffer_size), X.A.A.A.A.c_str(), X.A.A.B, X.A.B, X.B);
         
         if(required_size < buffer_size)
           {
@@ -299,7 +299,7 @@ namespace arma_boost
           buffer = new char[buffer_size];
           }
         
-        required_size = arma_snprintf(buffer, buffer_size, X.A.A.A.A.A.c_str(), X.A.A.A.B, X.A.A.B, X.A.B, X.B);
+        required_size = arma_snprintf(buffer, size_t(buffer_size), X.A.A.A.A.A.c_str(), X.A.A.A.B, X.A.A.B, X.A.B, X.B);
         
         if(required_size < buffer_size)
           {
@@ -351,7 +351,7 @@ namespace arma_boost
           buffer = new char[buffer_size];
           }
         
-        required_size = arma_snprintf(buffer, buffer_size, X.A.A.A.A.A.A.c_str(), X.A.A.A.A.B, X.A.A.A.B, X.A.A.B, X.A.B, X.B);
+        required_size = arma_snprintf(buffer, size_t(buffer_size), X.A.A.A.A.A.A.c_str(), X.A.A.A.A.B, X.A.A.A.B, X.A.A.B, X.A.B, X.B);
         
         if(required_size < buffer_size)
           {
@@ -403,7 +403,7 @@ namespace arma_boost
           buffer = new char[buffer_size];
           }
         
-        required_size = arma_snprintf(buffer, buffer_size, X.A.A.A.A.A.A.A.c_str(), X.A.A.A.A.A.B, X.A.A.A.A.B, X.A.A.A.B, X.A.A.B, X.A.B, X.B);
+        required_size = arma_snprintf(buffer, size_t(buffer_size), X.A.A.A.A.A.A.A.c_str(), X.A.A.A.A.A.B, X.A.A.A.A.B, X.A.A.A.B, X.A.A.B, X.A.B, X.B);
         
         if(required_size < buffer_size)
           {
